5-AI_Speech/requirements.txt
faster-whisper
torch
torchaudio
chromadb
pypdf2
sentence-transformers
transformers
accelerate
llama-cpp-python
piper-tts
sounddevice